feat: jpa credit card repository

+++ src/main/java/com/jmreisswitz/creditcards/domain/creditcard/CreditCardId.java
@@ -1,4 +1,4 @@
 package com.jmreisswitz.creditcards.domain.creditcard;
 
-public record CreditCardId(Long value) {
+public record CreditCardId(Integer value) {
 }
