+++ src/main/java/com/jmreisswitz/creditcards/infrastructure/controller/response/CardCreatedResponse.java
@@ -4,17 +4,17 @@ import java.io.Serializable;
 
 public class CardCreatedResponse implements Serializable {
 
-    private Long cardId;
+    private Integer cardId;
 
-    public CardCreatedResponse(Long cardId) {
+    public CardCreatedResponse(Integer cardId) {
         this.cardId = cardId;
     }
 
-    public Long getCardId() {
+    public Integer getCardId() {
         return cardId;
     }
 
-    public void setCardId(Long cardId) {
+    public void setCardId(Integer cardId) {
         this.cardId = cardId;
     }
 }
